+++ src/Documentation/composer.json
@@ -26,5 +26,8 @@
         "branch-alias": {
             "dev-main": "3.3.x-dev"
         }
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^11.2"
     }
 }
