+++ src/Symfony/Tests/Action/EntrypointActionTest.php
@@ -11,32 +11,22 @@
 
 declare(strict_types=1);
 
-namespace ApiPlatform\Tests\Action;
+namespace ApiPlatform\Symfony\Tests\Action;
 
-use ApiPlatform\Action\EntrypointAction;
-use ApiPlatform\Api\Entrypoint;
+use ApiPlatform\Documentation\Action\EntrypointAction;
+use ApiPlatform\Documentation\Entrypoint;
 use ApiPlatform\Metadata\Resource\Factory\ResourceNameCollectionFactoryInterface;
 use ApiPlatform\Metadata\Resource\ResourceNameCollection;
 use ApiPlatform\State\ProcessorInterface;
 use ApiPlatform\State\ProviderInterface;
 use PHPUnit\Framework\TestCase;
-use Prophecy\PhpUnit\ProphecyTrait;
+use Symfony\Component\HttpFoundation\Request;
 
 /**
  * @author Amrouche Hamza <hamza.simperfit@gmail.com>
  */
 class EntrypointActionTest extends TestCase
 {
-    use ProphecyTrait;
-
-    public function testGetEntrypoint(): void
-    {
-        $resourceNameCollectionFactoryProphecy = $this->prophesize(ResourceNameCollectionFactoryInterface::class);
-        $resourceNameCollectionFactoryProphecy->create()->willReturn(new ResourceNameCollection(['dummies']));
-        $entrypoint = new EntrypointAction($resourceNameCollectionFactoryProphecy->reveal());
-        $this->assertEquals(new Entrypoint(new ResourceNameCollection(['dummies'])), $entrypoint());
-    }
-
     public function testGetEntrypointWithProviderProcessor(): void
     {
         $expected = new Entrypoint(new ResourceNameCollection(['dummies']));
@@ -47,6 +37,6 @@ class EntrypointActionTest extends TestCase
         $processor = $this->createMock(ProcessorInterface::class);
         $processor->expects($this->once())->method('process')->willReturnArgument(0);
         $entrypoint = new EntrypointAction($resourceNameCollectionFactory, $provider, $processor);
-        $this->assertEquals($expected, $entrypoint());
+        $this->assertEquals($expected, $entrypoint(Request::create('/')));
     }
 }
