+++ tests/Doctrine/EventListener/PurgeHttpCacheListenerTest.php
@@ -14,9 +14,9 @@ declare(strict_types=1);
 namespace ApiPlatform\Tests\Doctrine\EventListener;
 
 use ApiPlatform\Doctrine\EventListener\PurgeHttpCacheListener;
-use ApiPlatform\Exception\InvalidArgumentException;
-use ApiPlatform\Exception\ItemNotFoundException;
 use ApiPlatform\HttpCache\PurgerInterface;
+use ApiPlatform\Metadata\Exception\InvalidArgumentException;
+use ApiPlatform\Metadata\Exception\ItemNotFoundException;
 use ApiPlatform\Metadata\GetCollection;
 use ApiPlatform\Metadata\IriConverterInterface;
 use ApiPlatform\Metadata\ResourceClassResolverInterface;
